requirements.txt
click==6.6
Flask==0.11.1
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pluggy==0.3.1
py==1.4.31
pytest==2.9.2
python-magic==0.4.12
tox==2.3.1
virtualenv==15.0.2
Werkzeug==0.11.10
